refactor: NumberApplicatorToCar 제거, CarRacingRequestDto 제거
