+++ src/main/java/racingcar/domain/racing/CarRacing.java
@@ -4,22 +4,21 @@ package racingcar.domain.racing;
 import java.util.List;
 import racingcar.domain.Car;
 import racingcar.domain.Cars;
-import racingcar.dto.CarRacingRequestDto;
 import racingcar.numbergenerator.NumberGenerator;
+import racingcar.view.CarRacingInputs;
 
 public class CarRacing {
     private final Cars cars;
-    private final NumberApplicatorToCar numberApplicatorToCar;
+    private final NumberGenerator numberGenerator;
     private final int totalRacingTryTime;
     private int finishedRacingTime;
 
-    public CarRacing(CarRacingRequestDto carRacingRequestDto, NumberGenerator numberGenerator) {
+    public CarRacing(CarRacingInputs carRacingInputs, NumberGenerator numberGenerator) {
         this.cars = new Cars();
-        this.numberApplicatorToCar
-            = new NumberApplicatorToCar(cars, numberGenerator);
-        this.totalRacingTryTime = carRacingRequestDto.getRacingTryTime();
+        this.numberGenerator = numberGenerator;
+        this.totalRacingTryTime = carRacingInputs.getRacingTryTime();
         this.finishedRacingTime = 0;
-        setCarsByCarNamesBeforeRacing(carRacingRequestDto.getCarNames());
+        setCarsByCarNamesBeforeRacing(carRacingInputs.getCarNames());
     }
 
     private void setCarsByCarNamesBeforeRacing(List<String> carNames) {
@@ -27,8 +26,16 @@ public class CarRacing {
     }
 
     public void raceOneTime() {
-        numberApplicatorToCar.apply();
+        applyGeneratedNumbersToCars();
         finishedRacingTime++;
+    }
+
+    private void applyGeneratedNumbersToCars() {
+        List<Car> allCars = cars.getAllCars();
+        for (Car car : allCars) {
+            int generatedCar = numberGenerator.generate();
+            car.goForwardDependsOnNumber(generatedCar);
+        }
     }
 
     public boolean isEnd() {
