+++ src/test/java/racingcar/domain/racing/utils/CarsCreatorTestUtils.java
@@ -2,7 +2,6 @@ package racingcar.domain.racing.utils;
 
 import java.util.ArrayList;
 import java.util.List;
-import racingcar.domain.Car;
 
 public class CarsCreatorTestUtils {
     public static List<String> createCarNamesSizeOf(int carSize) {
@@ -11,14 +10,5 @@ public class CarsCreatorTestUtils {
             carNames.add(String.valueOf(i));
         }
         return carNames;
-    }
-
-    public static List<Car> createCarsSizeOf(int carSize) {
-        List<String> carNames = createCarNamesSizeOf(carSize);
-        List<Car> cars = new ArrayList<>();
-        for (String carName : carNames) {
-            cars.add(new Car(carName));
-        }
-        return cars;
     }
 }
